requirements.txt
futures==3.0.5
oauthlib==2.0.0
pgnparser==1.0
Pillow==3.4.2
pycparser==2.17
python-chess==0.15.4
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
